+++ src/components/Profile/ProfileHeader/ProfileHeader.module.scss
@@ -42,9 +42,7 @@
   }
 
   .profile-setting-button {
-    img {
-      cursor: pointer;
-    }
+    cursor: pointer;
 
     button {
       width: 94px;
